% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmplot.R
\name{kmplot}
\alias{kmplot}
\title{kmplot}
\usage{
kmplot(data = NULL, plotType = "typical")
}
\arguments{
\item{data}{The data should be the clean dat2 from dataprocess.}

\item{plotType}{This argument decides which style ("simple", "specified", "typical".}
}
\value{
There is a image for kmplot.
}
\description{
kmplot for TCGA
}
\details{
By default, for this function you can draw the kmplot.
}
\examples{
\dontrun{
## Plot the kmplot automatically
kmplot()

## Plot kmfit using "specified" style
kmplot(plotType = "specified")
}
}
\author{
Wei Zhou <247328181@qq.com>
}
