% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMPPcui.R
\name{KMPPcui}
\alias{KMPPcui}
\title{KMPPcui}
\usage{
KMPPcui()
}
\description{
Launch the KMPPc user interface in local machine
}
\details{
This function will automatically launch the KMPPc user interface in a web browser. 
The user interface provides many powerful functions which is not available by command line programming.
It also provides a much easier and more convenient way to quickly explore TCGA data and construct kmplot you are interested in.
The user interface can also be accessed by https://zhouwei.shinyapps.io/KMPPc/. Neither R nor any packages are required in this online version.
However, it is highly recommended that the user interface be launched locally for faster running speed.
}
\examples{
\dontrun{
   KMPPcui()
}
}
\author{
Wei Zhou <247328181@qq.com>
}
