% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataexplore.R
\name{dataexplore}
\alias{dataexplore}
\title{dataexplore}
\usage{
dataexplore(cancerType, studyId = NULL, dataType = NULL)
}
\arguments{
\item{cancerType}{The cancertype you'd like to explore, including "BreastCancer", "PancreaticCancer", "Glioma"}

\item{studyId}{This argument you can explore by yourself, because there is too many researches or database you can take advantage of.}

\item{dataType}{It provides "mrna", "CNA", "mutation" for more information to dig.}
}
\value{
There is a RData file (survival_inputdata.Rdata) exported to the data folder.
}
\description{
dataexplore for TCGA
}
\details{
By default, this function first input cancerType such as "BreastCancer" to explore the dataset whatever you like.
}
\examples{
\dontrun{
## Explore the interested dataset by input the specified cancertype
dataexplore(cancerType = "BreastCancer", dataType = "mrna")

## If you have interested datatype, you can input, such as mrna, CNA, mutation, etc
dataexplore(cancerType = "BreastCancer", studyId = NULL, dataType = "mrna")
}
}
\author{
Wei Zhou <247328181@qq.com>
}
