% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataprocess.R
\name{dataprocess}
\alias{dataprocess}
\title{dataprocess}
\usage{
dataprocess(clinicaldata = NULL, exprSet = NULL, x_axis = NULL, y_axis = NULL)
}
\arguments{
\item{clinicaldata}{The data can be loaded by the results of dataexplore. You can load("survival_inputdata.Rdata") to get both of datasets.}

\item{exprSet}{The data can be loaded by the results of dataexplore. You can load("survival_inputdata.Rdata") to get both of datasets.}

\item{x_axis}{This argument is the x axis for time.}

\item{y_axis}{This argument is the y axis for status.}
}
\value{
There is a clean data after dataprocess.
}
\description{
dataprocess for TCGA
}
\details{
By default, this function you can filter some dirty data to prepare for kmplot.
}
\examples{
\dontrun{
## Plot the kmplot automatically
kmplot()

## Plot kmfit using "specified" style
kmplot(plotType = "specified")
}
}
\author{
Wei Zhou <247328181@qq.com>
}
